requirements.txt
requests
validators
beautifulsoup4
python-dotenv
fuzzywuzzy
python-Levenshtein
